requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-bidi==0.6.7
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.5.10
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
